allow one Get to each inbox at a time
This should cancel old long polling connection when you request the same inbox
for new message with another context.

one thing I noticed when I put the server behind cloudflare is that if a long
polled get request is cancelled from my browser by refreshing the page then
connecting again, the send message from the other peer won't reach me, and I
think this is because cloudflare doesn't cancel the request to inbox server or
another server in between like HAproxy that I have, so the solution here is to
always allow the newest request to get the message and cancel any old requests.

That will be handy in different cases, that to be honest I still don't know
about